refactor : added all filter function in context
